+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 2.2.0 - TBD
-- Added parsers for A2OJ and HDU Online Judge
+- Added parsers for A2OJ, HDU Online Judge, Toph, PEG Judge, POJ, omegaUp and COJ
 - Fixed the HackerRank and the SPOJ parsers
 
 # 2.1.1 - June 2nd, 2018